Renaming of aggregate rankings to combined rankings
