+++ combined_cs2_rankings/optimal_score.md
