+++ combined_cs2_rankings/output/.gitignore
