+++ combined_cs2_rankings/teamname_mapping.csv
